refactor(app): import from index.js instead of constants
